Begin the development of cult slang.
Despite all the arguments about "the best programming language,"
any developer will spend most of ta's time in the shell.

The shell is the linguistic equivalent of slang, vernacular, or
everyday speech. One can argue all day about the benefits of French
vs German vs Latin and shit, but at the end of the day, you'll most
likely relax back into the everyday informal jargon of your native
language (most likely English, Mandarin, or Spanish, if you live
on earth these days).

Therefore, one's shell config file is the proper location for the
definitions and implementations of the slang one uses in one's
everyday life.

This commit adds zshrc, which will be the home for the private
slang of an old white bunny, and his bird-y cultwife.