Add bin/cult-install script.
This script will install the cult's config files by making symlinks
from the locations the shell expects to find them (like ~/.zshrc)
to the repo (wherever that happens to be on your machine).

Pass the repo's relative or absolute path as an argument to the
script in order to install the cult.